language_model: Remove dependency on `inline_completion_button`
