+++ crates/inline_completion_button/Cargo.toml
@@ -23,7 +23,6 @@ paths.workspace = true
 settings.workspace = true
 supermaven.workspace = true
 ui.workspace = true
-util.workspace = true
 workspace.workspace = true
 zed_actions.workspace = true
