+++ crates/language_model/Cargo.toml
@@ -32,7 +32,6 @@ futures.workspace = true
 google_ai = { workspace = true, features = ["schemars"] }
 gpui.workspace = true
 http_client.workspace = true
-inline_completion_button.workspace = true
 log.workspace = true
 menu.workspace = true
 ollama = { workspace = true, features = ["schemars"] }
