+++ crates/language_model/src/provider/copilot_chat.rs
@@ -383,9 +383,7 @@ impl Render for ConfigurationView {
                                         .icon_size(IconSize::Medium)
                                         .style(ui::ButtonStyle::Filled)
                                         .full_width()
-                                        .on_click(|_, cx| {
-                                            inline_completion_button::initiate_sign_in(cx)
-                                        }),
+                                        .on_click(|_, cx| copilot::initiate_sign_in(cx)),
                                 )
                                 .child(
                                     div().flex().w_full().items_center().child(
